+++ src/main/java/org/lessons/java/gamebase/Service/GenreService.java
@@ -3,7 +3,9 @@ package org.lessons.java.gamebase.Service;
 import java.util.List;
 import java.util.Optional;
 
+import org.lessons.java.gamebase.Model.Game;
 import org.lessons.java.gamebase.Model.Genre;
+import org.lessons.java.gamebase.Repository.GameRepository;
 import org.lessons.java.gamebase.Repository.GenreRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,9 @@ public class GenreService {
     
     @Autowired
     private GenreRepository repo;
+
+    @Autowired
+    private GameRepository gameRepo;
 
     public List<Genre> getAll(){
         return repo.findAll();
@@ -24,8 +29,23 @@ public class GenreService {
         return attempt;
     };
 
-    public void deleteById(Integer id){
-        repo.deleteById(id);
+    public Boolean deleteById(Integer id){
+
+        Genre genre = getById(id).isPresent() ? getById(id).get() : null; 
+
+        if (genre != null) {
+            for (Game game : genre.getGames()) {
+                game.getGenres().remove(genre);
+            }
+            
+            repo.deleteById(id);
+
+            return true;
+        }
+
+        return false;
+
+
     }
 
     public void save(Genre genre){
